Remove common channel (#14)
A 'common' channel was a "catch all" channel for all of the public
repositories. We've disabled it a while ago, because we don't use it
anymore and it only generated more costs. We don't intend to turn it on
again, so it's safe to just remove it from the code.