Simplify channels logic (#16)
API uses only private user channels at this point, so we can safely
always send only to private user channels. Additionaly, sending updates
to public repos should be done only when a repo is public and when we're
not forcing private channels.

This simplifies the code and also it simplifies further work on merging
both platforms.